# Conflicts:
#	TiJian/TiJian.xcodeproj/project.pbxproj
#	TiJian/TiJian/CommonModels/ProductModel.h
#	TiJian/TiJian/CommonViews/GshopCarTableViewCell.m
#	TiJian/TiJian/ViewControllers/个人中心/PersonalCenterController.m
#	TiJian/TiJian/ViewControllers/商城/GStoreHomeViewController.m

+++ TiJian/TiJian/HelpTools/Categories/UIView+Additions.h
@@ -17,7 +17,7 @@
  *  @param selector slector
  *  @param tag      包含button tag值
  */
-- (void)addTaget:(id)target action:(SEL)selector tag:(int)tag;
+- (void)addTaget:(id)target action:(SEL)selector tag:(NSInteger)tag;
 /**
  *  给view加圆角
  *
